fix : 낙관적 락 Version 값 Int -> Long으로 변경

+++ src/main/kotlin/com/hhplus/domain/entity/Booking.kt
@@ -30,5 +30,5 @@ final class Booking(seatId : Int, bookingDate : LocalDateTime, status : BookingS
     var userUuid : Long ?= null
 
     @Version
-    private val version : Int ?= null
+    private val version : Long ?= null
 }
